Create html
In this application, we start with a population of 1000, which is displayed in an HTML span element with an id of "population". The two buttons on the page, labeled "Add Person" and "Remove Person", call JavaScript functions addPerson() and removePerson() when clicked.

In these functions, we update the population variable and the text inside the populationSpan element to reflect the new population count. We also make sure that the removePerson() function doesn't allow the population to go below zero.